wire koin di and register nav entries per feature
The app could not start: MainActivity resolved MainAppViewModel through
koinViewModel() but Koin was never started, and NavDisplay's
entryProvider was empty so nothing rendered.

- start Koin from SmsmanApplication with androidContext + feature modules
- each feature owns its route registration via Koin's navigation<T> {}
  DSL; koinEntryProvider assembles them for NavDisplay
- add placeholder screens so every route resolves
- widen AppRouteKey.Chat.chatId to Long to match Telephony thread ids
- pass onBack explicitly so pops share the size-guarded back() helper

Verified on device: launches to Home, no crash, back exits to launcher.

+++ app/src/main/java/com/abrarshakhi/smsman/common/navigation/AppRouteKey.kt
@@ -13,7 +13,7 @@ sealed interface AppRouteKey : NavKey {
 
     @Serializable
     data class Chat(
-        val chatId: Int // Replace with chat ID which will be better.
+        val chatId: Long,
     ) : AppRouteKey
 
     @Serializable
